Moved TimeDomainEquationSystemProblemOperator into its own file.
